+++ R/tbl_regression.R
@@ -25,15 +25,6 @@
 #' models use `broom.mixed::tidy(x, effects = "fixed")`. Specify `tidy_fun = broom.mixed::tidy`
 #' to print the random components.
 #'
-#' @section Note:
-#' The N reported in the output is the number of observations
-#' in the data frame `model.frame(x)`. Depending on the model input, this N
-#' may represent different quantities. In most cases, it is the number of people or
-#' units in your model.  Here are some common exceptions.
-#' 1. Survival regression models including time dependent covariates.
-#' 2. Random- or mixed-effects regression models with clustered data.
-#' 3. GEE regression models with clustered data.
-#'
 #' This list is not exhaustive, and care should be taken for each number reported.
 #' @param x Regression model object
 #' @param exponentiate Logical indicating whether to exponentiate the