+++ NEWS.md
@@ -1,5 +1,15 @@
 # gtsummary (development version)
 
+* Updates to `tbl_regression()` documentation. The model N is no longer reported by default, and removed that section from the help file. (#998)
+
+* Added function `tbl_butcher()` to reduce the size of a gtsummary table. After an object has been butchered, other gtsummary functions may not be able to execute on the object.
+
+* Added new function `tidy_robust()` that will add robust standard errors, confidence intervals, and p-values with `tbl_regression()` and `tbl_uvregression()`. The function is a wrapper for `parameters::model_paramters()`. (#979)
+
+* Migrated a new function `separate_p_footnotes()` from the {bstfun} package. Function allows users to separate the composite footnote listing the tests performed in `add_p()`, and replaces it with individual footnotes for each test name.
+
+* Migrated a new function `tbl_split()` from the {bstfun} package. Function allows users to split a {gtsummary} table into multiple tables.
+
 * Deleted deprecated functions `add_comparison()`, `add_global()`, `tab_style_bold_p()`, `tab_style_bold_labels()`, `tab_style_italicize_levels()`, `tab_style_italicize_labels()`, `tab_style_bold_levels()`.
 
 * The following deprecated arguments have been removed: `tbl_summary(group=)`, `as_gt(omit=)`.